Refactor FileList model and update services
Replace the FileList class with a record and add a Mime type property for better file handling. Update service classes - FileService and Common classes to accommodate the changes.

Update the FileService to fetch the Mime type while listing the file details to support proper file handling. A GetMimeType method is added in the Common class to get the Mime type for each file based on their extension. Added Mime type property in the FileListItem record constructor to hold the Mime type value fetched. Added extensive Mime type list in Common class for mapping, and a default fallback if mapping not found.