Add Console testing project
Added a new Console testing project to the solution. 'CMD_Test' project was created to allow running different commands on different OS shells and view their outputs. This will facilitate command output testing in Linux and Windows environments. New Dockerfile is also added to create an environment for the test.