Add multiple assemblies for implemented interfaces and AgentService API
This commit includes updates to support multiple assemblies for implemented interfaces in the Shared.Common.cs file. The GetImplementedInterfaceOf method now accepts an array of assemblies, which provides greater flexibility for interface implementations. This change improves modularity and adaptability of the code for interface invoking and handling.

This commit also introduces the foundational structure of the AgentService API. This includes the addition of new files for the API itself (including appsettings, Dockerfile, and Program.cs), along with associated projects for the Domain, Application, and Infra layers. These additions lay the groundwork necessary to expand the AgentService API functionality in future work.

Newly added PluginCollection and IPluginInformation are part of Plugin mechanism that enables dynamic functionality extension.

Additionally, a few changes were made in project organisation including file relocation and renaming to make the structure more coherent and maintainable.