Refactored HTTP mapping methods to improve fluidity
The mapping methods in HttpHandlerExt.cs were refactored to return 'IEndpointRouteBuilder' instead of 'RouteHandlerBuilder'. This change allows the routing setup code to be more fluid and intuitive. This refactoring also removes the need for storing the 'group' variable in FileEndpoints.cs, we can directly chain mapping operations on the route builder. This gets rid of some intermediate variables and makes the code cleaner and easier to follow.