Remove Dockerfile and Program.cs in Hardware_Info module
Deleted files Dockerfile and Program.cs in the Hardware_Info module as they are no longer required. The functionality provided by these components has been shifted to another module to streamline the operation. These files were responsible for setting up a docker environment and collecting hardware info, but this can now be accomplished in a more efficient way with the new module. This change helps to eliminate redundancy and enhance maintainability.