Add Program.cs and Dockerfile for Dashboard Agent
The Project.cs file was renamed to Agent.UI.csproj and a new Dockerfile was added to enable running the agent in docker. New services were registered in the Program.cs file for the RazorPages, Blazor Server, WeatherForecastService. Also, exception handling and HTTPS redirection were implemented. Default project file OutputType was removed and TargetFramework was reduced to net7.0. Handlers for BlazorHub and fallback page were created. This change was made to allow the agent to support deployment through Docker and server-side rendering of Blazor.