Add toast notification component and Boxicons font
Added Toast notification component for alerting users about certain actions or events in the system. Created related classes ToastType and ToastPlacement for further customization of the toast. Injected the toast into the Index.razor page and implemented a test usage on a button click. Refactored paths in main.js and _Host.cshtml files. Added Boxicons font in _Host.cshtml and modified path in boxicons.css for proper loading of icons.