Add file manager feature to the application
This commit includes changes made to add the functionality of a file manager to the application. These changes consist of service setup in ConfigureService.cs and the addition of necessary interfaces and models to establish this feature. Also, the user interface for this functionality has been introduced with FileManager.razor page. The footer page has been updated with the new links. Changes made to Dashboard/Index.razor and _Host.cshtml are for a better UI experience. The Infrastructure was set up to handle API requests related to the file manager. Also, unit test cases were updated for this new functionality. This feature will enable users to manage files within the application.