Add image handling functionality to the Docker connectors
This commit introduces a new functionality to Docker connectors to handle images. A 'Common' project was added to the solution, a reference to which is included in the Docker SSH connector. A couple of parser helpers were created to parse raw image lists and translate them into an Image object. The Image model was updated to match properties from both API and SSH connectors responses, removing unused properties, and adding necessary ones. Raw Image models for both connectors were created as well. 'note.txt' was included to the solution items. This functionality will streamline handling Docker images across our project.