Add SSH and API Docker connector test projects
Added project references for Docker.Connectors.SSH.Test and Docker.Connectors.API.Test to the solution, along with necessary configuration for each in Debug and Release builds. These changes are required to include unit testing capabilities for Docker connectors in the solution.

In addition to that, created new test class ImagesListParserTests in both projects to have a dedicated place for testing the ImagesListParser functionality.

Also updated the ImagesListParser from being instantiable to a static class, as it only contains static methods and shouldn't be instantiated for usage.
Finally, added usage of Xunit through a global usings declaration file in both test projects for better clarity and less repetition in the code.