Add models, helpers, and tests for parsing raw container data
Added the 'ContainerRaw' and 'ContainersListParser' models for the process of extracting and structuring details from raw Docker container data. Additionally, the 'Host' and 'NetworkConfig' models were created to represent the network settings of each container. Adjustments were also been made to the 'Port' model. These changes will enable the Docker connector API to more accurately and effectively parse and work with the data it receives. To verify the parsing process, tests have been added in 'ContainersListParserTests'.