Change visibility of classes and methods to internal
Visibility of certain classes in API Helpers and Models was changed to 'internal'. The methods in the 'ImagesListParser' and 'ContainersListParser' classes were also changed to 'internal'. This change is intended to limit the access of these classes and methods to within the assembly, thereby improving encapsulation. The 'InternalsVisibleToAttribute' was added to allow visibility for unit testing.