Add Logging.Abstractions package to Common project
The 'Microsoft.Extensions.Logging.Abstractions' package was added to the Common.csproj file. This allows logging to be done in a platform-agnostic way, which provides more flexibility when using different logging frameworks or storing logs in different locations. The addition of this package would improve debuggability and error tracking for the project.