Refactored authentication to include endpoint details
In previous versions, the endpoint was managed separately from user authentication. This lead to redundancies and lack of cohesion when creating an ApiConnector instance. This change integrates the endpoint into the authentication system in order to construct a DockerClientConfiguration object directly, which simplifies the logic and makes the API more intuitive to use. Usernames and passwords remain unchanged and will continue to work as before.