Add support for retrieving Docker networks
Several changes were made to the API and SSH connectors to retrieve Docker networks. An additional namespace import was added to the API connector, and a new 'GetNetworks' function was implemented in both connectors. This function retrieves network information using the Docker.Net library for the API connector, and an SSH command for the SSH connector. In addition, the IConnector interface was updated to include the 'GetNetworks' a command generator for listing networks was added to the SSH connector. These changes provide access to Docker network information, enhancing overall functionality.