Add user's real name to ClaimsIdentity and display in dashboard
The NameIdentifier and UserName were previously the only claims included in the ClaimsIdentity. Now, GivenName is also included to provide more valuable user information. Also, the real name of the current user is fetched from the claims and displayed at the dashboard. This makes the UI more user friendly and personalized.

Additionally, redirect logic was added to the login and dashboard layout pages to make sure users are navigated correctly based on their authentication status.