Refactor services registration in Program.cs
Services registration was moved from Program.cs into a new dedicated class ConfigureServices in the Dashboard namespace. This is part of a bigger refactoring to improve application architecture by having a separate class responsible for services registration, increasing maintainability and separation of concerns. The Connectors.Docker.ConfigureServices assembly was also added to the registered services.