Add show/hide password functionality in Login page
The code implements a feature to toggle visibility of the password field on the Login page. The password input type changes based on the state of the _showPassword variable. An onclick event handler is added to toggle this state, changing the visibility of the entered password and the icon displayed beside the field.