Add Guid Id property and update UI in Dashboard page
A Guid Id property has been added to the code part of the DashboardPage.razor file. This will be useful for identifying each dashboard page uniquely. Also, the UI has been significantly updated in the Main.razor file, replacing simple buttons with a dropdown menu in the ToolbarSection for better user experience and functionality.