Update encryption and decryption to use byte array key
The code is refactored in CredentialManager for encryption and decryption to use byte array keys instead of a key file. The Encrypt and Decrypt methods in ICredentialManager and their corresponding tests in CredentialManagerTest were updated to reflect these changes. This provides more flexibility as keys don't need to be managed as separate files.