Add Dashboard.Application.Test project and initial test cases
This commit includes the introduction of the Dashboard.Application.Test project into the solution. It includes initial test cases for handling the addition of different types of credentials. As part of the setup, Xunit has been globally included and the necessary references and tools for conducting the tests have been added into the project file.