Add IDateTimeProvider and IUserProvider interfaces
This commit adds the IDateTimeProvider interface with a Now property and extends IUserProvider with a GetCurrentUser method. These interfaces will allow us to more easily manage datetime and user data throughout the application. The User class constructor has also been refactored to accept an IdColumn type as the id parameter.