Update GetAsync method to return list of credentials
The GetAsync method in the ICredentialRepository interface and the CredentialRepository class was modified to return a list of credentials instead of a single one. This change allows better handling of situations where multiple credentials match a name query. A corresponding method was also added to the BaseVaultRepository to facilitate this lookup.