Update method return types in repository classes
This commit updates the GetAsync and GetAllAsync methods in the ICredentialRepository and BaseVaultRepository classes. These methods were previously returning a list of nullable Credentials (potentially including null values), which has now been refactored to return a list of non-nullable Credentials instead, thus omitting null potentials. The change is propagated in the CredentialRepository as well.