Add User page to Dashboard
User page has been added to the dashboard for better user management. A unique User ID is also added, for accurate identification and data management of users. An 'if' condition was included in the Index.razor file to render the User component when the current ID matches the User ID.