Update AuthorizerType enum and refine toolbar UI
AuthorizerType enum is updated to standardize the naming of LDAP type. Toolbar UI in Users page is also enhanced for a more organized look and feel, with improved functionality. Now, it shows the selected provider type and offers user a dropdown menu to select between Local and LDAP. Moreover, additional context-sensitive options have been added for LDAP provider type and a new Settings icon is incorporated.