Update entity model and create initial migration
This commit introduces changes to the ApplicationDbContextModelSnapshot by updating the entity model, changing some property types to Guid and adjusting relationships. This required the creation of an initial migration script to align the database with the updated entity model rule. The update in the entity model will open the door to more robust and clear data handling within the application.