Remove "Authorizer.Local" project and update code dependencies
This commit removes the "Authorizer.Local" project and updates file paths, namespaces and method signatures to reflect this change. Items that were previously within the "Authorizer.Local" scope have been relocated or removed accordingly. The changes also include relevant updates on the solution and project files, and rename actions on class files and moving them to the "Dashboard" project.