Update user management in Dashboard's Authentication
In the Users section of the dashboard's authentication, various changes have been implemented. New features include status-specific (active or inactive) user action options and the ability to set a password for local users. In addition, a new method, `AddUser()`, has been added to enable addition of a new user from the dashboard interface.