Add password setting functionality to users page
Added a "Set Password" functionality in the Users.razor page, creating an option to set a password for a user in the drop-down menu for each user. This involved introducing a new SetPassword.razor component and a new SetPassword method. The new component includes validation and a user feedback system.