Add update user status functionality to dashboard
This commit adds the functionality for updating the status of a user in the dashboard. The changes include new command, command validator, and command handler classes for updating user status, along with modifying the UI to display and modify the user status. Also, the code for deleting a user has been removed.